+++ src/circuit_library.h
@@ -39,7 +39,7 @@ public:
     size_t num_circuits() const {return m_circuits.size();}
     CircuitDescription *circuit_by_idx(size_t idx) const;
     CircuitDescription *circuit_by_name(const char *name) const;
-    size_t circuit_idx(CircuitDescription *circuit) const;
+    uint32_t circuit_idx(CircuitDescription *circuit) const;
     void clear_circuits();
 
     // references
